final-ish beach tweak, tested worldgen times
not 100% happy with the beaches, but they are still looking pretty good!

also implemented multithreading in biome generation, specifically for the two floodfills. it saves ~3 seconds for a 64x64 world.

the test results were encouraging! in release builds, Biome_Generator::generate() takes 3-4 seconds. bulk of worldgen time still seems to be the grass automaton - move that to perlin noise as well, eventually. (or find another noise generation algorithm!)

for the full set of results, as well as a lovely little chart, check google sheets.