fixed machine behavior wrt subclasses
the only odd thing is that items from conflicting reactions may be added as long as no reaction is active. not going to fuss about it for now, though.

something nice would be a "max reagant count" dictating the number of slots appearing in the reaction panel, but that's not *super* important.

another "would-be-nice" is a reaction panel showing all available reactions.