added tooltips, condensed reaction panels
fine placement hasn't been figured out yet so they are simply a jumbled mess of sprites and words... for now. in addition, the frame is still invisible

the inventory interface has turned into a disgusting spaghetti monolith, and i am sure i will be punished for it in the afterlife

rapidly approaching 0.03! :)