animated buildings
all of them are animated which is of course a huge waste, but i'll see what the performance impact is like and adjust if needed. besides, it'll be useful for things like doors.

also added a new "Tile_Map" using directive, which will be implemented next. it is set to an unordered_map keyed to a vector2i, which should improve performance.